requirements.txt
fastapi
uvicorn[standard]
pydantic
playwright
requests
python-dotenv
pandas
matplotlib
beautifulsoup4
openai
